Auto stash before merge of "feature/srat" and "origin/feature/edward"

+++ Backend/src/main/java/com/example/demo/Repository/UserRepository.java
@@ -4,6 +4,10 @@ import com.example.demo.Entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import javax.persistence.EntityManager;
+
 @Repository
-public interface UserRepository extends JpaRepository<User,Long> {
+//public interface UserRepository extends JpaRepository<User,Long> {
+//}
+public interface UserRepository extends JpaRepository<User,Integer> {
 }
